staging: sync: Remove fence names by default
Copying fence debug names for frequent users of sync fences causes a
measurably significant portion of CPU time to be spent on the strlcpy.
Additionally, the sync_fence struct is allocated very frequently, so
wasting space on an unneeded 64-byte name is suboptimal.

Remove the 64-byte name from the struct to reduce its size considerably
and eliminate unnecessary operations on fence names.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[shagbag913: ported to 3.18]
Signed-off-by: shagbag913 <sh4gbag913@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>